pdf-microservice/requirements.txt
# SyncLedger PDF Extraction Microservice
# Python FastAPI service for extracting invoice data from PDFs

# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# AWS SDK
boto3==1.34.14
aioboto3==12.2.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
orjson==3.9.10

# Date parsing
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Type hints
types-python-dateutil==2.8.19.14
